refactor register
