+++ .gitignore
@@ -2,4 +2,4 @@
 *.pyc
 site-log.log
 /__pycache__
-
+*.log
