+++ templates/login.html
@@ -11,9 +11,9 @@
         <div class="tema">
         <div class="int">
         <form method="POST">
-        <br/>Usuário:
+        <br/>Email:
 
-        <br/><input type="text" name="username"/> 
+        <br/><input type="text" name="email"/> 
         <br/>Senha:
         <br/><input type="password" name="password"/> <br/>
