+++ templates/register.html
@@ -1,8 +1,8 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="pt-br">
 <head>
     <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Register</title>
 </head>
@@ -15,8 +15,8 @@
         <div class="int">
             <form method="POST">
 
-            <br/>Nome de Usuário:
-            <br/><input type="text" name="username"/> <br/>
+            <br/>Email:
+            <br/><input type="text" name="email"/> <br/>
             
             <br/>Senha:
             <br/><input type="password" name="password"/> <br/>
@@ -24,13 +24,21 @@
             <br/>Confirmar Senha:
             <br/><input type="password" name="password_c"/> <br/>
             
+            <br/>Rua:
+            <br/><input type="text" name="street"/> <br/>
+
+            <br/>Numero:
+            <br/><input type="text" name="number"/> <br/>
+
+            <br/>Bairro:
+            <br/><input type="text" name="nbh"/> <br/>
+
             <br/>Cidade:
             <br/><input type="text" name="city"/> <br/>
 
             <br/>Estado:
             <br/><input type="text" name="state"/> <br/>
             
-
             <br/> <input type="submit" value="Entrar"/>
             </form>
         </div>
